pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openev-project"
version = "1.0.0"
description = "AI Security Code Reviewer — Unified AppSec Agent for OpenEnv & HF Spaces"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "pydantic>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.24.0",
    "openai>=1.0.0",
    "python-dotenv",
    "pyyaml",
    "gradio>=4.26.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Entry point for the unified app
app = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["server"]
py-modules = ["models", "inference", "client"]